requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain-mistralai
langchain-pinecone
langchain-core
langchain-text-splitters
python-dotenv
pinecone-client
pydantic
pydantic-settings
loguru
pytest
pytest-asyncio
pytest-cov
pytest-mock
httpx
faker